Separate model and schema of test data
